+++ src/pricefeeds/CanonicalPriceFeed.sol
@@ -65,7 +65,7 @@ contract CanonicalPriceFeed is OperatorStaking, SimplePriceFeed, CanonicalRegist
         OperatorStaking(
             AssetInterface(ofStakingAsset), stakingInfo[0], stakingInfo[1], stakingInfo[2]
         )
-        SimplePriceFeed(this, ofQuoteAsset, 0x0)
+        SimplePriceFeed(address(this), ofQuoteAsset, address(0))
     {
         registerAsset(
             ofQuoteAsset,
@@ -102,20 +102,18 @@ contract CanonicalPriceFeed is OperatorStaking, SimplePriceFeed, CanonicalRegist
 
     /// @dev Burn state for a pricefeed operator
     /// @param user Address of pricefeed operator to burn the stake from
-    /// @param amount Amount of stake to burn
+    /// @param burnAmount Amount of stake to burn
     /// @param data Additional data
-    function burnStake(address user, uint amount, bytes data)
+    function burnStake(address user, uint burnAmount, bytes data)
         external
         auth
     {
-        require(totalStakedFor(user) >= amount);
-        uint preStake = totalStakedFor(user);
-        uint postStake = sub(preStake, amount);
-        require(postStake >= minimumStake || postStake == 0);
-        updateCheckpointAtNow(stakesFor[user], amount, true);
-        updateCheckpointAtNow(stakeHistory, amount, true);
+        require(stakedAmounts[user] >= burnAmount);
+        uint preStake = stakedAmounts[user];
+        uint postStake = sub(preStake, burnAmount);
+        stakedAmounts[user] -= burnAmount;
         updateStakerRanking(user);
-        StakeBurned(user, amount, data);
+        emit StakeBurned(user, burnAmount, data);
     }
 
     // PUBLIC METHODS
@@ -131,7 +129,7 @@ contract CanonicalPriceFeed is OperatorStaking, SimplePriceFeed, CanonicalRegist
     {
         OperatorStaking.stake(amount, data);
     }
-  
+
     // function stakeFor(
     //     address user,
     //     uint amount,
@@ -157,8 +155,19 @@ contract CanonicalPriceFeed is OperatorStaking, SimplePriceFeed, CanonicalRegist
     /// @param ofAssets list of asset addresses
     function collectAndUpdate(address[] ofAssets)
         public
-        auth 
+        auth
         pre_cond(updatesAreAllowed)
+    {
+        uint[] memory newPrices = pricesToCommit(ofAssets);
+        priceHistory.push(
+            HistoricalPrices({assets: ofAssets, prices: newPrices, timestamp: block.timestamp})
+        );
+        _updatePrices(ofAssets, newPrices);
+    }
+
+    function pricesToCommit(address[] ofAssets)
+        view
+        returns (uint[])
     {
         address[] memory operators = getOperators();
         uint[] memory newPrices = new uint[](ofAssets.length);
@@ -174,10 +183,7 @@ contract CanonicalPriceFeed is OperatorStaking, SimplePriceFeed, CanonicalRegist
             }
             newPrices[i] = medianize(assetPrices);
         }
-        priceHistory.push(
-            HistoricalPrices({assets: ofAssets, prices: newPrices, timestamp: block.timestamp})
-        );
-        _updatePrices(ofAssets, newPrices);
+        return newPrices;
     }
 
     /// @dev from MakerDao medianizer contract
@@ -370,12 +376,15 @@ contract CanonicalPriceFeed is OperatorStaking, SimplePriceFeed, CanonicalRegist
         view
         returns(address[])
     {
-        address[] memory ofPriceFeeds = new address[](stakeRanking.length);
-        for (uint i; i < stakeRanking.length; i++) {
-            StakingPriceFeed stakingFeed = StakingPriceFeed(stakeRanking[i].staker);
+        address[] memory ofPriceFeeds = new address[](numStakers);
+        if (numStakers == 0) return ofPriceFeeds;
+        uint current = stakeNodes[0].next;
+        for (uint i; i < numStakers; i++) {
+            StakingPriceFeed stakingFeed = StakingPriceFeed(stakeNodes[current].data.staker);
             if (stakingFeed.owner() == _owner) {
                 ofPriceFeeds[i] = address(stakingFeed);
             }
+            current = stakeNodes[current].next;
         }
         return ofPriceFeeds;
     }